requirements.txt
##############################
# Core Python / Utils
##############################
python-dotenv>=1.0.1
tqdm>=4.66.1
requests>=2.32.0
pandas>=2.2.2
numpy>=1.26.0
Pillow>=10.3.0

##############################
# Neo4j Graph DB
##############################
neo4j>=5.19.0

##############################
# Vector Stores
##############################
chromadb>=0.5.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

##############################
# LLM / LangChain / LangGraph
##############################
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-experimental>=0.3.0
langgraph>=0.2.5

openai>=1.49.0

##############################
# HuggingFace Transformers
##############################
torch>=2.2.0
transformers>=4.44.0
accelerate>=0.30.1

##############################
# Vision Models (BLIP2)
##############################
git+https://github.com/huggingface/transformers.git
salesforce-lavis>=1.0.0

##############################
# ESMFold (local structure prediction)
##############################
fair-esm>=2.0.0

##############################
# Protein Biology Tools
##############################
biopython>=1.83

##############################
# FastAPI Backend
##############################
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

##############################
# Streamlit Frontend
##############################
streamlit>=1.37.0
py3Dmol>=2.0.4

##############################
# Additional Utilities
##############################
html2text>=2024.2.26
pydantic>=2.7.0
httpx>=0.27.0
ijson>=3.2.0
